Add fix-license command
It detects obsolete license specifications used by StrongLoop, and
converts them to a currently valid specification.

It also rewrites the LICENSE or LICENSE.md file for a valid license
specification, guaranteeing consistent text for all license files.

The existence of the tool should help keep current packages consistent,
as well as allowing any future changes we decide to make to the
licensing (such as adjustments to the format of the license files) much
easier, perhaps trivial.