change blip script to not depend on node
No matter how foolproof the JS is and how explicitly it tries to always
exit with 0, none of that matters if node isn't in the path for some
reason.

To handle this, add a simple `|| exit 0` to the end of the script entry
in package.json.

Also make sure we "upgrade" the previous style if we see it being used.