add argument forwarding support to slt-stage
Any arguments that are given are passed along to 'npm publish' when it
is run as the last step. This allows the publish to be done against a
different registry than the install, which is required when using a
restricted proxy.

Example usage:

```
export npm_config_registry=http://readonly/
slt-stage --registry=http://read-write/
```