slt-release: remove use of X.x-latest branches
Don't bother creating and updating the major-release tracking branches
that slt-release previously created. They aren't actually useful or
meaningful and have only caused confusion when people see them.

They were originally inteaded as a sign-post indicating what the most
recent release tag of a given major version was, since that may or may
not be the HEAD of a given branch, despite the HEAD of a given branch
having the same version specified in its package.json.