slt-release: commit shrinkwrap, but don't merge it
If there is an npm-shrinkwrap.json file present AND that file has not
been committed to the current branch, then commit it as part of the
release tag but filter it out of the actual merge back in to the base
branch.

Note that this mode is NOT triggered if npm-shrinkwrap.json is already
part of git. If you want to enable that style of development going
forward, you will have to commit it separately _before_ running
`slt-release`, otherwise it will be ignored and filtered out.

This is done using some slightly low-level git operations to manually
compose a merge commit in a way that neither 'git merge' or 'git commit'
interfaces allow. At least it is well commented!