slt-release: fallback to package.json for version
If no version or increment is given, use the version that is in
package.json for the release.

After the version has been identified, by argument, increment, or
package.json, THEN validate it for suitability.

The version must be:
 - unreleased on the registry being used
 - untagged in the git repository
 - a valid SemVer version string

If all of these criteria are met, then proceed to cut the release.